Refactored BungeeComm to ProxyLocalCommunicationManager
