+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/commands/IgnoreCommand.java
@@ -4,10 +4,10 @@ import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Command;
-import xyz.olivermartin.multichat.bungee.BungeeComm;
 import xyz.olivermartin.multichat.bungee.ChatControl;
 import xyz.olivermartin.multichat.bungee.ConfigManager;
 import xyz.olivermartin.multichat.bungee.MessageManager;
+import xyz.olivermartin.multichat.proxy.common.ProxyLocalCommunicationManager;
 
 public class IgnoreCommand extends Command {
 
@@ -50,7 +50,7 @@ public class IgnoreCommand extends Command {
 						MessageManager.sendSpecialMessage(sender, "ignore_unignored", target.getName());
 					}
 					
-					BungeeComm.sendIgnoreMap(((ProxiedPlayer) sender).getServer().getInfo());
+					ProxyLocalCommunicationManager.sendIgnoreMap(((ProxiedPlayer) sender).getServer().getInfo());
 
 				} else {
